bin/ghar: ignore .git/ and .gitignore
Currently all the entries which name starts with .git* are ignored. This
means you can't version files which name starts with .git* (global
.gitconfig for example).

Tweak the regular expression to only ignore entries with the name .git
or .gitignore. This will allow .gitconfig to be installable to $HOME.

Based on work by Kami