Recurses existing directories (but symlinks wrong)
Goal: implement symlinking for _individual files_ inside directories
which already exist.  For example, `~/.vim/filetype.vim` should be
symlinked from the repository, but the `~/.vim` directory should not.
I've got the recursive processing down... now I just need to fix the
constructor for `Link` so the target is relative to the `ghar` base.