"Fix" directly linked folders.
Don't try to make the basename of a link relative if the link is not a
collection, as the result is always the path ".". This was broken by #7.
This isn't an ideal fix as it doesn't handle the case where the target
already exists; it would be nice for it to merge recursively the way it
does for collections, but it's better than it failing entirely.